+++ ceammc/CHANGELOG.md
@@ -22,6 +22,10 @@
     "outlets": ["audio", "control"...] instead of outlet number
 - props:
     - prop.set renamed to prop.join (with prop<- alias)
+- spat.pan4~:
+    - @dist property renamed to @radius
+    - positional args order reversed, now is: radius, angle
+    - list support added for easy connection with ui.polar
 
 ## [0.9]
 ### Added: